move setup_clibs.py
