+++ python/amici/setup_clibs.py
